requirements.txt
streamlit>=1.36
numpy>=1.26
scipy>=1.11
matplotlib>=3.8
pandas>=2.2
pyyaml>=6.0